fix(servers): clear C4305/C4309/C4018 truncation and signedness warnings.
Clamp BYTE trade counts correctly, null-terminate Garner names with '\0', and align IMP/weather comparisons.

+++ source/src/gameserver/Weather.cpp
@@ -33,7 +33,7 @@ void CWeather::RandLocation(SubMap* pMap) {
 
 	CSkillStateRecord* pEff = GetCSkillStateRecordInfo(_btType);
 	if (pEff) {
-		if ((dwCurTick - g_dwLastWeatherTick[_btType]) > 1000 * g_Config.m_lWeather) // 每种天气, 每2分钟一次天气预报
+		if ((dwCurTick - g_dwLastWeatherTick[_btType]) > static_cast<DWORD>(1000 * g_Config.m_lWeather)) // 每种天气, 每2分钟一次天气预报
 		{
 			g_dwLastWeatherTick[_btType] = dwCurTick;
 			char szText[128];
